requirements.txt
Flask==3.1.1
Flask-Login
Flask-Mail
Flask-SQLAlchemy
python-dotenv
bcrypt
mercadopago
email-validator